# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/media": {
      "post": {
        "operationId": "MediaController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMediaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload a base64 image and get a media id back",
        "tags": [
          "media"
        ]
      }
    },
    "/api/media/{id}": {
      "get": {
        "operationId": "MediaController_metadata",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Media metadata",
        "tags": [
          "media"
        ]
      },
      "delete": {
        "operationId": "MediaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete your own media",
        "tags": [
          "media"
        ]
      }
    },
    "/api/media/{id}/raw": {
      "get": {
        "operationId": "MediaController_raw",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Original image bytes",
        "tags": [
          "media"
        ]
      }
    },
    "/api/media/{id}/thumbnail": {
      "get": {
        "operationId": "MediaController_thumbnail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Thumbnail bytes — use this in lists",
        "tags": [
          "media"
        ]
      }
    },
...
    "/api/invites/accept": {
      "post": {
        "operationId": "MembersController_acceptInvite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInviteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceMemberResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Redeem an invite code and join that place",
        "tags": [
          "members"
        ]
      }
    },
    "/api/places/{placeId}/members": {
      "get": {
        "operationId": "MembersController_list",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceMemberResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Everyone with access to this place",
        "tags": [
          "members"
        ]
      },
      "post": {
        "operationId": "MembersController_add",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceMemberResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add an accepted friend to this place",
        "tags": [
          "members"
        ]
      }
    },
    "/api/places/{placeId}/members/{memberId}": {
      "patch": {
        "operationId": "MembersController_updateRole",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceMemberResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change what someone can do in this place",
        "tags": [
          "members"
        ]
      },
      "delete": {
        "operationId": "MembersController_remove",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove someone from this place",
        "tags": [
          "members"
        ]
      }
    },
    "/api/places/{placeId}/leave": {
      "post": {
        "operationId": "MembersController_leave",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Leave a place you were added to",
        "tags": [
          "members"
        ]
      }
    },
    "/api/places/{placeId}/invites": {
      "post": {
        "operationId": "MembersController_createInvite",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceInviteResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an invite code for someone not on the app yet",
        "tags": [
          "members"
        ]
      },
      "get": {
        "operationId": "MembersController_listInvites",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceInviteResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invite codes that are still open",
        "tags": [
          "members"
        ]
      }
    },
    "/api/places/{placeId}/invites/{inviteId}": {
      "delete": {
        "operationId": "MembersController_revokeInvite",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "inviteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke an unused invite code",
        "tags": [
          "members"
        ]
      }
    },
...
  },
  "info": {
    "title": "FindMyStuff API",
    "description": "Remember where you kept your stuff.\n\nHierarchy: **Place** (home/office/locker) → **Storage** tree (room → almirah → shelf → box) → **Item**.\n\nEvery place-scoped route is nested under `/places/{placeId}` and is checked against your\nmembership of that place. Asking for a place you are not a member of returns **404**, not 403 —\nthe API never confirms that someone else's data exists.\n\nAll routes except signup, login, refresh, logout and health require\n`Authorization: Bearer <accessToken>`.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateMediaDto": {
        "type": "object",
        "properties": {
          "base64": {
            "type": "string",
            "description": "Image as base64. A data URI (\"data:image/jpeg;base64,...\") or raw base64 both work."
          },
          "label": {
            "type": "string",
            "description": "Optional caption/filename, stored for reference only"
          }
        },
        "required": [
          "base64"
        ]
      },
      "MediaResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Use this id when creating an item or storage"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "mimeType": {
            "type": "string",
            "example": "image/jpeg"
          },
          "sizeBytes": {
            "type": "number",
            "example": 284913
          },
          "width": {
            "type": "object",
            "nullable": true,
            "example": 1600
          },
          "height": {
            "type": "object",
            "nullable": true,
            "example": 1200
          },
          "checksum": {
            "type": "string",
            "description": "SHA-256 — re-uploading the same photo reuses the stored file"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ownerId",
          "mimeType",
          "sizeBytes",
          "width",
          "height",
          "checksum",
          "createdAt"
        ]
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Home — Demo deleted"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "UserCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Shristi"
          },
          "email": {
            "type": "string",
            "example": "shristi@example.com"
          },
          "avatarMediaId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "PlaceMemberResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER",
              "VIEWER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "INVITED",
              "ACTIVE",
              "REMOVED"
            ]
          },
          "joinedAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/UserCard"
          },
          "invitedBy": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserCard"
              }
            ]
          }
        },
        "required": [
          "id",
          "role",
          "status",
          "joinedAt",
          "user"
        ]
      },
...
      "AcceptInviteDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "FMS-INV-7K3QX2"
          }
        },
        "required": [
          "code"
        ]
      },
      "AddMemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "Must already be an accepted friend"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER",
              "VIEWER"
            ],
            "default": "MEMBER"
          }
        },
        "required": [
          "userId"
        ]
      },
      "UpdateMemberDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER",
              "VIEWER"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "CreateInviteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Who the invite is for — used to prefill and to notify"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER",
              "VIEWER"
            ],
            "default": "MEMBER"
          },
          "expiresInDays": {
            "type": "number",
            "default": 7,
            "minimum": 1,
            "maximum": 90
          }
        }
      },
      "PlaceInviteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "example": "INV-ZJ5Y0YAP",
            "description": "Share this code with the person"
          },
          "email": {
            "type": "object",
            "nullable": true,
            "example": "mum@example.com"
          },
          "phone": {
            "type": "object",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER",
              "VIEWER"
            ]
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "acceptedAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "code",
          "email",
          "phone",
          "role",
          "expiresAt",
          "acceptedAt"
        ]
      },
...
    }
  }
}